Simplify Composer installation instructions
Typing a command is simpler than editing a file: you get
auto-completion, there are less keys to press, you do not need to open
an editor… besides, "*" is not a really good version constraint,
because it does not protect the user from BC-breaks. Not specifying a
version constraint in the command will result in Composer applying
whatever the current best practice is.
This also removes outdated instructions for installing Composer, linking
to the dedicated page instead, because we probably do not want to
maintain this.