Fixed #236 - problem with automatic casting of numeric values
The automatic casting of numeric values in the BaseContext class
is limited by the PHP_INT_MAX constant. Any string with an numeric
value larger than PHP_INT_MAX gets automatically casted to an
integer with the value of PHP_INT_MAX.

This could lead to problems, for example when sending a http-header
with a large numeric value.

This workaround checks if the numeric value is equal or larger than
PHP_INT_MAX and does not cast to integer in this case.

Also added a test for this case.